requirements.txt
dash
python-dotenv
numpy==1.26.4
pandas
nltk
h5py==3.10.0
tensorflow-cpu==2.20.0
keras==3.4.1
gunicorn